% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_three}
\alias{add_three}
\title{Add three numbers}
\usage{
add_three(arg1, arg2, arg3)
}
\arguments{
\item{arg1}{First integer.}

\item{arg2}{Second integer.}

\item{arg3}{Third integer.}
}
\value{
The sum of three integers.
}
\description{
This functions add three integers.
}
\examples{
add_three(10, 11, 12)
}
