% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_two}
\alias{add_two}
\title{Add two numbers}
\usage{
add_two(arg1, arg2)
}
\arguments{
\item{arg1}{First integer.}

\item{arg2}{Second integer.}
}
\value{
The sum of two integers.
}
\description{
This functions add two integers.
}
\examples{
add_two(10, 11)
}
