% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{generate_Fmat}
\alias{generate_Fmat}
\title{Test importing by package}
\usage{
generate_Fmat(ntip)
}
\arguments{
\item{ntip}{Number of tips}
}
\value{
F-matrix.
}
\description{
This function generate a tree using rcoal() and return the Fmat
}
\examples{
generate_Fmat(10)
}
