% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{has_root}
\alias{has_root}
\title{Test importing by functions}
\usage{
has_root(ntip)
}
\arguments{
\item{ntip}{Number of tips}
}
\value{
If the tree is rooted.
}
\description{
This function generate a tree using rcoal() and return if the tree is rooted
}
\examples{
has_root(10)
}
