% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{multiply_three}
\alias{multiply_three}
\title{Multiply three numbers}
\usage{
multiply_three(arg1, arg2, arg3)
}
\arguments{
\item{arg1}{First integer.}

\item{arg2}{Second integer.}

\item{arg3}{Third integer.}
}
\value{
The product of three integers.
}
\description{
This function mutipies three integers.
}
\examples{
multiply_three(10, 11, 12)
}
