% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{multiply_two}
\alias{multiply_two}
\title{Multiply two numbers}
\usage{
multiply_two(arg1, arg2)
}
\arguments{
\item{arg1}{First integer.}

\item{arg2}{Second integer.}
}
\value{
The product of two integers.
}
\description{
This function mutipies two integers.
}
\examples{
multiply_two(10, 11)
}
