% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{my_function}
\alias{my_function}
\title{My Function Title}
\usage{
my_function(arg1, arg2 = "default_value")
}
\arguments{
\item{arg1}{Description of the first argument.}

\item{arg2}{Description of the second argument, including its default value.}
}
\value{
Description of the value returned by the function.
}
\description{
A more detailed description of what \code{my_function} does.
}
\examples{
my_function(10, "hello")
my_function(5)
}
